test: extract media migration snapshot fixture
